날씨, wifi ui 수정 중
