+++ Interface/JSY8869/threads.ui
@@ -6,8 +6,8 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>800</width>
-    <height>269</height>
+    <width>964</width>
+    <height>572</height>
    </rect>
   </property>
   <property name="windowTitle">
@@ -21,28 +21,59 @@
        <string>Testing threads</string>
       </property>
       <layout class="QGridLayout" name="gridLayout_3">
-       <item row="0" column="0">
+       <item row="3" column="0">
+        <widget class="QWidget" name="widget" native="true">
+         <widget class="QPushButton" name="wifi_button">
+          <property name="geometry">
+           <rect>
+            <x>690</x>
+            <y>0</y>
+            <width>221</width>
+            <height>161</height>
+           </rect>
+          </property>
+          <property name="text">
+           <string>wifi 설정</string>
+          </property>
+         </widget>
+         <widget class="QPushButton" name="weather_button">
+          <property name="geometry">
+           <rect>
+            <x>0</x>
+            <y>0</y>
+            <width>221</width>
+            <height>161</height>
+           </rect>
+          </property>
+          <property name="text">
+           <string>오늘의 날씨는?</string>
+          </property>
+         </widget>
+         <widget class="QTextEdit" name="textEdit">
+          <property name="geometry">
+           <rect>
+            <x>220</x>
+            <y>0</y>
+            <width>471</width>
+            <height>161</height>
+           </rect>
+          </property>
+          <property name="html">
+           <string>&lt;!DOCTYPE HTML PUBLIC &quot;-//W3C//DTD HTML 4.0//EN&quot; &quot;http://www.w3.org/TR/REC-html40/strict.dtd&quot;&gt;
+&lt;html&gt;&lt;head&gt;&lt;meta name=&quot;qrichtext&quot; content=&quot;1&quot; /&gt;&lt;style type=&quot;text/css&quot;&gt;
+p, li { white-space: pre-wrap; }
+&lt;/style&gt;&lt;/head&gt;&lt;body style=&quot; font-family:'Gulim'; font-size:9pt; font-weight:400; font-style:normal;&quot;&gt;
+&lt;p align=&quot;justify&quot; style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
+&lt;p align=&quot;justify&quot; style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px; font-size:20pt; font-weight:600;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
+&lt;p align=&quot;justify&quot; style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;span style=&quot; font-size:20pt; font-weight:600;&quot;&gt;오늘 서울의 날씨는 맑습니다.&lt;/span&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
+          </property>
+         </widget>
+        </widget>
+       </item>
+       <item row="1" column="0" rowspan="2">
         <layout class="QGridLayout" name="gridLayout">
-         <item row="0" column="2">
-          <widget class="QProgressBar" name="progressBar">
-           <property name="value">
-            <number>24</number>
-           </property>
-          </widget>
-         </item>
-         <item row="2" column="2">
-          <widget class="QProgressBar" name="progressBar_3">
-           <property name="value">
-            <number>24</number>
-           </property>
-          </widget>
-         </item>
-         <item row="1" column="2">
-          <widget class="QProgressBar" name="progressBar_2">
-           <property name="value">
-            <number>24</number>
-           </property>
-          </widget>
+         <item row="0" column="3" rowspan="3">
+          <layout class="QVBoxLayout" name="verticalLayout"/>
          </item>
          <item row="0" column="0">
           <widget class="QPushButton" name="pushButton">
@@ -54,9 +85,6 @@
            </property>
           </widget>
          </item>
-         <item row="0" column="3" rowspan="3">
-          <layout class="QVBoxLayout" name="verticalLayout"/>
-         </item>
          <item row="1" column="0">
           <widget class="QPushButton" name="pushButton_2">
            <property name="styleSheet">
@@ -67,13 +95,20 @@
            </property>
           </widget>
          </item>
-         <item row="2" column="0">
-          <widget class="QPushButton" name="pushButton_3">
+         <item row="1" column="2">
+          <widget class="QProgressBar" name="progressBar_2">
+           <property name="value">
+            <number>24</number>
+           </property>
+          </widget>
+         </item>
+         <item row="1" column="1">
+          <widget class="QPushButton" name="pushButton_5">
            <property name="styleSheet">
-            <string notr="true">background-color: rgb(85, 255, 0);</string>
+            <string notr="true">background-color: rgb(255, 85, 127);</string>
            </property>
            <property name="text">
-            <string>Start Thread 3</string>
+            <string>Stop Thread 2</string>
            </property>
           </widget>
          </item>
@@ -87,13 +122,27 @@
            </property>
           </widget>
          </item>
-         <item row="1" column="1">
-          <widget class="QPushButton" name="pushButton_5">
+         <item row="2" column="2">
+          <widget class="QProgressBar" name="progressBar_3">
+           <property name="value">
+            <number>24</number>
+           </property>
+          </widget>
+         </item>
+         <item row="2" column="0">
+          <widget class="QPushButton" name="pushButton_3">
            <property name="styleSheet">
-            <string notr="true">background-color: rgb(255, 85, 127);</string>
+            <string notr="true">background-color: rgb(85, 255, 0);</string>
            </property>
            <property name="text">
-            <string>Stop Thread 2</string>
+            <string>Start Thread 3</string>
+           </property>
+          </widget>
+         </item>
+         <item row="0" column="2">
+          <widget class="QProgressBar" name="progressBar">
+           <property name="value">
+            <number>24</number>
            </property>
           </widget>
          </item>
